requirements-test.txt
robotframework>=7.0,<8.0
pyserial>=3.5,<4.0
